chore(querier): Refactor the store and querier interface.
