+++ pkg/compactor/deletion/delete_requests_store.go
@@ -233,7 +233,6 @@ func (ds *deleteRequestsStore) GetCacheGenerationNumber(ctx context.Context, use
 		}
 		return false
 	})
-
 	if err != nil {
 		return "", err
 	}
