+++ pkg/querier/http_test.go
@@ -51,33 +51,6 @@ func TestInstantQueryHandler(t *testing.T) {
 	})
 }
 
-func TestTailHandler(t *testing.T) {
-	defaultLimits := defaultLimitsTestConfig()
-	limits, err := validation.NewOverrides(defaultLimits, nil)
-	require.NoError(t, err)
-
-	api := NewQuerierAPI(mockQuerierConfig(), nil, limits, log.NewNopLogger())
-
-	req, err := http.NewRequest("GET", `/`, nil)
-	require.NoError(t, err)
-	q := req.URL.Query()
-	q.Add("query", `{app="loki"}`)
-	req.URL.RawQuery = q.Encode()
-	err = req.ParseForm()
-	require.NoError(t, err)
-
-	ctx := user.InjectOrgID(req.Context(), "1|2")
-	req = req.WithContext(ctx)
-	require.NoError(t, err)
-
-	rr := httptest.NewRecorder()
-	handler := http.HandlerFunc(api.TailHandler)
-
-	handler.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusBadRequest, rr.Code)
-	require.Equal(t, "multiple org IDs present", rr.Body.String())
-}
-
 type slowConnectionSimulator struct {
 	sleepFor   time.Duration
 	deadline   time.Duration
@@ -88,7 +61,6 @@ func (s *slowConnectionSimulator) ServeHTTP(_ http.ResponseWriter, r *http.Reque
 	ctx := r.Context()
 	if err := ctx.Err(); err != nil {
 		panic(fmt.Sprintf("context already errored: %s", err))
-
 	}
 	time.Sleep(s.sleepFor)
 
@@ -221,6 +193,7 @@ func TestSeriesHandler(t *testing.T) {
 		require.JSONEq(t, expected, res.Body.String())
 	})
 }
+
 func TestVolumeHandler(t *testing.T) {
 	ret := &logproto.VolumeResponse{
 		Volumes: []logproto.Volume{
