+++ pkg/querier/querier_mock_test.go
@@ -7,12 +7,7 @@ import (
 	"math"
 	"time"
 
-	"github.com/grafana/loki/v3/pkg/logql/log"
-	"github.com/grafana/loki/v3/pkg/logql/syntax"
-
 	"github.com/grafana/loki/pkg/push"
-
-	"github.com/grafana/loki/v3/pkg/loghttp"
 
 	"github.com/grafana/dskit/grpcclient"
 	"github.com/grafana/dskit/ring"
@@ -24,14 +19,18 @@ import (
 	"google.golang.org/grpc/health/grpc_health_v1"
 	grpc_metadata "google.golang.org/grpc/metadata"
 
-	logql_log "github.com/grafana/loki/v3/pkg/logql/log"
-
+	"github.com/grafana/loki/v3/pkg/compactor/deletion"
 	"github.com/grafana/loki/v3/pkg/distributor/clientpool"
 	"github.com/grafana/loki/v3/pkg/ingester/client"
 	"github.com/grafana/loki/v3/pkg/iter"
+	"github.com/grafana/loki/v3/pkg/loghttp"
 	"github.com/grafana/loki/v3/pkg/logproto"
 	"github.com/grafana/loki/v3/pkg/logql"
+	"github.com/grafana/loki/v3/pkg/logql/log"
+	logql_log "github.com/grafana/loki/v3/pkg/logql/log"
+	"github.com/grafana/loki/v3/pkg/logql/syntax"
 	"github.com/grafana/loki/v3/pkg/logqlmodel"
+	"github.com/grafana/loki/v3/pkg/querier/pattern"
 	"github.com/grafana/loki/v3/pkg/storage/chunk"
 	"github.com/grafana/loki/v3/pkg/storage/chunk/fetcher"
 	"github.com/grafana/loki/v3/pkg/storage/config"
@@ -157,7 +156,7 @@ func (f mockIngesterClientFactory) newIngesterClientMockFactory(c *querierClient
 
 // newIngesterClientMockFactory creates a factory function always returning
 // the input querierClientMock
-func newIngesterClientMockFactory(c *querierClientMock) ring_client.PoolFactory {
+func newIngesterClientMockFactory(c ring_client.PoolClient) ring_client.PoolFactory {
 	return ring_client.PoolAddrFunc(func(_ string) (ring_client.PoolClient, error) {
 		return c, nil
 	})
@@ -262,62 +261,6 @@ func (c *querySampleClientMock) RecvMsg(_ interface{}) error {
 
 func (c *querySampleClientMock) Context() context.Context {
 	return context.Background()
-}
-
-// tailClientMock is mockable version of Querier_TailClient
-type tailClientMock struct {
-	util.ExtendedMock
-	logproto.Querier_TailClient
-	recvTrigger chan time.Time
-}
-
-func newTailClientMock() *tailClientMock {
-	return &tailClientMock{
-		recvTrigger: make(chan time.Time, 10),
-	}
-}
-
-func (c *tailClientMock) Recv() (*logproto.TailResponse, error) {
-	args := c.Called()
-	return args.Get(0).(*logproto.TailResponse), args.Error(1)
-}
-
-func (c *tailClientMock) Header() (grpc_metadata.MD, error) {
-	return nil, nil
-}
-
-func (c *tailClientMock) Trailer() grpc_metadata.MD {
-	return nil
-}
-
-func (c *tailClientMock) CloseSend() error {
-	return nil
-}
-
-func (c *tailClientMock) Context() context.Context {
-	return context.Background()
-}
-
-func (c *tailClientMock) SendMsg(_ interface{}) error {
-	return nil
-}
-
-func (c *tailClientMock) RecvMsg(_ interface{}) error {
-	return nil
-}
-
-func (c *tailClientMock) mockRecvWithTrigger(response *logproto.TailResponse) *tailClientMock {
-	c.On("Recv").WaitUntil(c.recvTrigger).Return(response, nil)
-
-	return c
-}
-
-// triggerRecv triggers the Recv() mock to return from the next invocation
-// or from the current invocation if was already called and waiting for the
-// trigger. This method works if and only if the Recv() has been mocked with
-// mockRecvWithTrigger().
-func (c *tailClientMock) triggerRecv() {
-	c.recvTrigger <- time.Now()
 }
 
 // storeMock is a mockable version of Loki's storage, used in querier unit tests
@@ -655,8 +598,8 @@ func mockLogfmtStreamWithLabels(_ int, quantity int, lbls string) logproto.Strea
 		streamLabels = labels.EmptyLabels()
 	}
 
-	lblBuilder := logql_log.NewBaseLabelsBuilder().ForLabels(streamLabels, streamLabels.Hash())
-	logFmtParser := logql_log.NewLogfmtParser(false, false)
+	lblBuilder := log.NewBaseLabelsBuilder().ForLabels(streamLabels, streamLabels.Hash())
+	logFmtParser := log.NewLogfmtParser(false, false)
 
 	// used for detected fields queries which are always BACKWARD
 	for i := quantity; i > 0; i-- {
@@ -765,9 +708,9 @@ func (q *querierMock) Series(ctx context.Context, req *logproto.SeriesRequest) (
 	return args.Get(0).(func() *logproto.SeriesResponse)(), args.Error(1)
 }
 
-func (q *querierMock) Tail(_ context.Context, _ *logproto.TailRequest, _ bool) (*Tailer, error) {
-	return nil, errors.New("querierMock.Tail() has not been mocked")
-}
+// func (q *querierMock) Tail(_ context.Context, _ *logproto.TailRequest, _ bool) (*Tailer, error) {
+// 	return nil, errors.New("querierMock.Tail() has not been mocked")
+// }
 
 func (q *querierMock) IndexStats(_ context.Context, _ *loghttp.RangeQuery) (*stats.Stats, error) {
 	return nil, nil
@@ -833,7 +776,7 @@ func (q *querierMock) DetectedLabels(ctx context.Context, req *logproto.Detected
 	return resp.(*logproto.DetectedLabelsResponse), err
 }
 
-func (q *querierMock) WithPatternQuerier(_ PatterQuerier) {}
+func (q *querierMock) WithPatternQuerier(_ pattern.PatterQuerier) {}
 
 type engineMock struct {
 	util.ExtendedMock
@@ -869,4 +812,14 @@ func (tl mockTenantLimits) TenantLimits(userID string) *validation.Limits {
 
 func (tl mockTenantLimits) AllByUserID() map[string]*validation.Limits {
 	return tl
+}
+
+type mockDeleteGettter struct {
+	user    string
+	results []deletion.DeleteRequest
+}
+
+func (d *mockDeleteGettter) GetAllDeleteRequestsForUser(_ context.Context, userID string) ([]deletion.DeleteRequest, error) {
+	d.user = userID
+	return d.results, nil
 }
