--- pkg/querier/tail_mock_test.go
@@ -1,10 +1,0 @@
-package querier
-
-import "github.com/grafana/loki/v3/pkg/logproto"
-
-func mockTailResponse(stream logproto.Stream) *logproto.TailResponse {
-	return &logproto.TailResponse{
-		Stream:         &stream,
-		DroppedStreams: []*logproto.DroppedStream{},
-	}
-}
